# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Multi-tenant CRM SaaS API (Phase 1 foundation)
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/delete-account:
    delete:
      operationId: deleteUserAccount
      tags: [auth]
      summary: Permanently delete the authenticated user's account
      description: >
        Requires the authenticated user to type DELETE. The external Clerk
        user, every organization where they currently hold an owner role, all
        memberships, and their local identity are permanently removed. Records
        in organizations they do not own remain and user references are
        detached.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserAccountDeletionConfirmation"
      responses:
        "200":
          description: Account deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserAccountDeletionResponse"
        "400":
          description: Confirmation must be exactly DELETE
        "401":
          description: Not authenticated
        "404":
          description: User not found
        "409":
          description: Account deletion is already in progress
        "502":
          description: Clerk deletion failed; retry is safe
        "500":
          description: The relational purge failed and was rolled back; retry is safe
...
  /orgs/{orgId}:
    get:
      operationId: getOrg
      tags: [orgs]
      summary: Org details (requires membership)
      parameters:
        - $ref: "#/components/parameters/OrgId"
      responses:
        "200":
          description: Org
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "403":
          description: Not a member
    patch:
      operationId: updateOrg
      tags: [orgs]
      summary: Update org name/plan (owner/admin)
      parameters:
        - $ref: "#/components/parameters/OrgId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationUpdate"
      responses:
        "200":
          description: Updated org
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
    delete:
      operationId: deleteOrganization
      tags: [orgs]
      summary: Permanently delete an organization (owner only)
      description: Requires the authenticated organization owner to type DELETE. All organization data and private objects are permanently removed; global user accounts are preserved.
      parameters:
        - $ref: "#/components/parameters/OrgId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationDeletionConfirmation"
      responses:
        "200":
          description: Organization deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationDeletionResponse"
        "400":
          description: Confirmation must be exactly DELETE
        "401":
          description: Not authenticated
        "403":
          description: Only the organization owner can delete
        "404":
          description: Organization not found
        "409":
          description: A deletion is already in progress
        "502":
          description: An external deletion dependency failed; retry is safe
        "500":
          description: The relational purge failed and was rolled back; retry is safe
  /organizations/{orgId}:
    delete:
      operationId: deleteOrganizationAlias
      tags: [orgs]
      summary: Permanently delete an organization (owner-only alias)
      parameters:
        - $ref: "#/components/parameters/OrgId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationDeletionConfirmation"
      responses:
        "200":
          description: Organization deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationDeletionResponse"
        "400":
          description: Confirmation must be exactly DELETE
        "401":
          description: Not authenticated
        "403":
          description: Only the organization owner can delete
        "404":
          description: Organization not found
        "409":
          description: A deletion is already in progress
        "502":
          description: An external deletion dependency failed; retry is safe
        "500":
          description: The relational purge failed and was rolled back; retry is safe
...
components:
  parameters:
    OrgId:
      name: orgId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    Organization:
      type: object
      required: [id, name, slug, plan, enabledFeatures]
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        plan:
          type: string
          enum: [essential, professional, enterprise, custom]
        enabledFeatures:
          type: array
          items: { type: string }
        stripeCustomerId: { type: ["string", "null"] }
        stripeSubscriptionId: { type: ["string", "null"] }
        createdAt: { type: ["string", "null"] }
    OrganizationUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        plan:
          type: string
          enum: [essential, professional, enterprise, custom]
    OrganizationDeletionConfirmation:
      type: object
      additionalProperties: false
      required: [confirmation]
      properties:
        confirmation: { type: string, enum: [DELETE] }
    OrganizationDeletionResponse:
      type: object
      additionalProperties: false
      required: [success]
      properties:
        success: { type: boolean, enum: [true] }
    UserAccountDeletionConfirmation:
      type: object
      additionalProperties: false
      required: [confirmation]
      properties:
        confirmation: { type: string, enum: [DELETE] }
    UserAccountDeletionResponse:
      type: object
      additionalProperties: false
      required: [success]
      properties:
        success: { type: boolean, enum: [true] }
...