# lib/api-zod/src/generated/types/userAccountDeletionConfirmation.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Multi-tenant CRM SaaS API (Phase 1 foundation)
 * OpenAPI spec version: 0.1.0
 */
import type { UserAccountDeletionConfirmationConfirmation } from './userAccountDeletionConfirmationConfirmation';

export interface UserAccountDeletionConfirmation {
  confirmation: UserAccountDeletionConfirmationConfirmation;
}