# lib/api-zod/src/generated/types/userAccountDeletionResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Multi-tenant CRM SaaS API (Phase 1 foundation)
 * OpenAPI spec version: 0.1.0
 */

export interface UserAccountDeletionResponse {
  success: true;
}